reorganization phase 1
